create a function hyperbola and import module_b from init

+++ useful_package/module_b.py
@@ -1,2 +1,2 @@
 def hyperbola(x):
-    return 3/x
+    return float(3/x)
